move _build_dataloaders to ptychography_base
